Add files via upload
Added all by me introduced functions in the beginning and also all imports. Display of Lat/Lon is now much better (given in °S, ... instead of - and +), but only if using my method for South America plots (format_sa_plot(...)). For Altiplano it should also be usable but you will need to specify the optional argument used_extents= when calling the function.